+++ packages/@overeng/megarepo/nix/build.nix
@@ -14,7 +14,7 @@ let
     packageDir = "packages/@overeng/megarepo";
     workspaceRoot = src;
     pnpmDepsHash = "sha256-x+YpNp1tAi3GBBNDKF4wInvQ/yUPIPKBAOmJl2DBAEg=";
-    lockfileHash = "sha256-LwFRtt/hPEdieGD7zA7urtlf5wWidUKMfiN12hD3iJ0=";
+    lockfileHash = "sha256-qyq1kUDmfBOxK4em1Tk0D1OH1v34xudAmrYkyU5DKWI=";
     packageJsonDepsHash = "sha256-Jg/6gay8Kv9EYZ+Cvi6rku8PMmptthh5Sv513W79Q9w=";
     smokeTestArgs = [ "--help" ];
     inherit gitRev commitTs dirty;